Inject SSH keys in async thread
