+++ plugins/plugin-machine/che-plugin-machine-ext-server/src/main/java/org/eclipse/che/ide/ext/machine/server/MachineModule.java
@@ -11,8 +11,8 @@
 package org.eclipse.che.ide.ext.machine.server;
 
 import com.google.inject.AbstractModule;
-import org.eclipse.che.ide.ext.machine.server.ssh.KeysInjector;
 import org.eclipse.che.ide.ext.machine.server.ssh.WorkspaceSshKeys;
+import org.eclipse.che.ide.ext.machine.server.ssh.WorkspaceStatusSubscriber;
 import org.eclipse.che.inject.DynaModule;
 
 @DynaModule
@@ -20,7 +20,7 @@ public class MachineModule extends AbstractModule {
 
   @Override
   protected void configure() {
-    bind(KeysInjector.class).asEagerSingleton();
+    bind(WorkspaceStatusSubscriber.class).asEagerSingleton();
     bind(WorkspaceSshKeys.class).asEagerSingleton();
   }
 }
