+++ plugins/plugin-machine/che-plugin-machine-ext-server/src/main/java/org/eclipse/che/ide/ext/machine/server/ssh/KeysInjector.java
@@ -15,9 +15,6 @@ import static org.eclipse.che.api.workspace.shared.Constants.SERVER_EXEC_AGENT_H
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
-import javax.annotation.PostConstruct;
-import javax.inject.Inject;
-import javax.inject.Singleton;
 import org.eclipse.che.agent.exec.client.ExecAgentClient;
 import org.eclipse.che.agent.exec.client.ExecAgentClientFactory;
 import org.eclipse.che.agent.exec.shared.dto.GetProcessResponseDto;
@@ -25,15 +22,11 @@ import org.eclipse.che.agent.exec.shared.dto.ProcessStartResponseDto;
 import org.eclipse.che.api.core.NotFoundException;
 import org.eclipse.che.api.core.ServerException;
 import org.eclipse.che.api.core.model.workspace.Runtime;
-import org.eclipse.che.api.core.model.workspace.WorkspaceStatus;
 import org.eclipse.che.api.core.model.workspace.runtime.Machine;
 import org.eclipse.che.api.core.model.workspace.runtime.Server;
-import org.eclipse.che.api.core.notification.EventService;
-import org.eclipse.che.api.core.notification.EventSubscriber;
 import org.eclipse.che.api.ssh.server.SshManager;
 import org.eclipse.che.api.ssh.server.model.impl.SshPairImpl;
 import org.eclipse.che.api.workspace.server.WorkspaceManager;
-import org.eclipse.che.api.workspace.shared.dto.event.WorkspaceStatusEvent;
 import org.eclipse.che.commons.env.EnvironmentContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -42,63 +35,44 @@ import org.slf4j.LoggerFactory;
  * Injects public parts of ssh keys in the machine after container start
  *
  * @author Sergii Leschenko
+ * @author Max Shaposhnyk
  */
-@Singleton // must be eager
 public class KeysInjector {
 
   private static final Logger LOG = LoggerFactory.getLogger(KeysInjector.class);
 
-  private final EventService eventService;
   private final SshManager sshManager;
   private final WorkspaceManager workspaceManager;
   private final ExecAgentClientFactory execAgentClientFactory;
 
-  @Inject
   public KeysInjector(
-      EventService eventService,
       SshManager sshManager,
       WorkspaceManager workspaceManager,
       ExecAgentClientFactory execAgentClientFactory) {
-    this.eventService = eventService;
     this.sshManager = sshManager;
     this.workspaceManager = workspaceManager;
     this.execAgentClientFactory = execAgentClientFactory;
   }
 
-  @PostConstruct
-  public void start() {
-    eventService.subscribe(
-        new EventSubscriber<WorkspaceStatusEvent>() {
-          @Override
-          public void onEvent(WorkspaceStatusEvent event) {
-            if (event.getStatus() != WorkspaceStatus.RUNNING) {
-              return;
-            }
+  public void injectPublicKeys(String workspaceId) {
+    final List<String> execServerUrls = getExecServerUrls(workspaceId);
+    if (execServerUrls.isEmpty()) {
+      return; // no exec servers installed, so not possible to execute commands
+    }
 
-            final String workspaceId = event.getWorkspaceId();
-            final List<String> execServerUrls = getExecServerUrls(workspaceId);
-            if (execServerUrls.isEmpty()) {
-              return; // no exec servers installed, so not possible to execute commands
-            }
+    List<String> publicKeys = getPublicKeys(workspaceId);
+    if (publicKeys.isEmpty()) {
+      return; // no keys found, exiting
+    }
 
-            List<String> publicKeys = getPublicKeys(workspaceId);
-            if (publicKeys.isEmpty()) {
-              return; // no keys found, exiting
-            }
-
-            StringBuilder commandLine = new StringBuilder("mkdir ~/.ssh/ -p");
-            for (String publicKey : publicKeys) {
-              commandLine
-                  .append(" && echo '")
-                  .append(publicKey)
-                  .append("' >> ~/.ssh/authorized_keys");
-            }
-            final String command = commandLine.toString();
-            for (String serverUrl : execServerUrls) {
-              doInjectPublicKeys(serverUrl, workspaceId, command);
-            }
-          }
-        });
+    StringBuilder commandLine = new StringBuilder("mkdir ~/.ssh/ -p");
+    for (String publicKey : publicKeys) {
+      commandLine.append(" && echo '").append(publicKey).append("' >> ~/.ssh/authorized_keys");
+    }
+    final String command = commandLine.toString();
+    for (String serverUrl : execServerUrls) {
+      doInjectPublicKeys(serverUrl, workspaceId, command);
+    }
   }
 
   private List<String> getExecServerUrls(String workspaceId) {
